Move project files into root of solution
As a web application project, we don't expect this solution to require multiple projects. As such, the standard organization of placing each project in its own folder only serves to bury nearly all code by one level. Given this, I've moved all code from the `/Ignia.Web` project directory into the root of the solution (which is also the root of the repository).