Migrate `ControllerFactory` to `IControllerActivator`
In ASP.NET Framework, the _composition root_ was a `DefaultControllerFactory` implementation. In ASP.NET Core, it's now an `IControllerActivator`. In addition, ASP.NET Core introduces a new format—view components—which can also be registered via an `IViewComponentActivator`. These can be combined into a joint composition root for both, which we place in the `IgniaActivator`.

This is registered in the `Startup` file, and effectively replaces not only the legacy `IgniaControllerFactory`, but also the `Global.asax` file, which was previously responsible for establishing the singleton lifestyle for the shared cache.