Introduce custom `ErrorController`
In OnTopic 3.x, there was a shared `ErrorController` as part of OnTopic. This wasn't a very elegant solution, and actually limited flexibility while centralizing very little actual code. The new implementation, based on the new OnTopic 4.4.0 routing extensions, allows standard pages to be used for error handling. This does require a number of actions to be added to the `ErrorController`, but eliminates the need for a custom `Error.cshtml`.