Move original scripts into precompilation location
In a previous commit, we moved all client script under `wwwroot` (e7d8fd4). That's because we didn't yet have a (production ready) build process established, and so those needed to be in the main library. In anticipation of that, custom scripts are now being moved (back) to `/Shared/Scripts`. In subsequent commits, we'll (reintroduce) a gulp build process for generating these files.