Updated dependencies in preparation for new build process
The updated dependencies are based on a) moving client-side resources to local resources, in preference over their CDNs, and b) setting up the foundation for our new Gulp build process, which uses different plugins than our original setup (which was never committed to the repository). In addition, this updates the libraries we were already downloading via npm to their latest versions.