Updated paths for client-side resources
With the new `GulpFile.js` build process in place (7f99af3) we can update the paths to point to the new locations. Generally, there are four differences:
1. We no longer use the `*.min.js` extension for our custom scripts, in preference for source maps,
2. Template specific scripts (e.g., `Library.js`) are now under `/Views`,
3. Template specific scripts are no longer consolidated (i.e., there is both `Library.js` and `Library.Animations.js`),
4. Third-party fonts, styles, and scripts are now loaded locally, instead of relying on CDNs.

The last item comes with tradeoffs. It potentially slows down the download since we're not relying on well-optimized external CDNs, which users may already have cached copies of. But it makes it easier for us maintain the latest version and set our own cache policies.