Delete third-party dependencies and generated files
Third-party dependencies are now downloaded via npm and distributed using the `GulpFile.js`. Local resources are now mastered in the `/Shared/` directory, and then compiled and published to the `/wwwroot/Shared/` directory, also via the `GulpFile.js`. As such, there is no need to maintain these files as part of the repository since they are dynamically generated.