Implemented cache buster tag helper across site
Implemented the `asp-append-version` tag helper to help prevent unwanted caching of local resources when they've been modified. This works by appending a hash in the query string based on the file contents. This was partially implemented for the main `Scripts.js` and `Style.css` in a previous commit (c34d297), but hadn't been rolled out across the board to, notably, the view-specific resources.