Ensured all scripts were (properly) deferred
Not all external script references were set to `defer`. And several of the ones that were weren't actually setup correctly, with both the `defer` and `async` properties set (these are orthogonal). This used to be preferred for backward compatibility with browsers that supported one but not the other, but is no longer needed—and, in fact, effectively treated the scripts as `async` in modern browsers, which isn't what we usually want.