Deferred call to `foundation()` initializer
In a previous commit, I deferred all JavaScript files (15bad26). I had missed, however, that there was one instance of an inline script which depended on two of these files—jQuery and Foundation. Inline scripts can't be marked with `defer` so they instead need to implement an event listener for the `DOMContentLoaded` event. This resolves that issue.