Introduced a new `InlineScript` section
This isn't technically needed right now, but if future templates want to inject scripts that are deferred, the `InlineScript` section makes that easy by injecting the JavaScript directly into the `DOMContentLoaded` event listener established in the previous commit (7233d25). This is a standard convention Ignia uses for all of our new sites, but which hadn't yet been established for Ignia's own site.