Updated connection string configuration instructions
The connection string is no longer configured via the `connectionStrings.config` file, but is instead configured via the `secrets.js` (at least locally). Further, there is no longer a membership database, as that was part of the legacy Forms Based Authentication, which has been replaced with OpenID Connect for Azure (8d1dc4a).