Remvoed Foundation scripts that aren't actively used
Previously, we were including all Foundation JavaScript via `foundation.min.js`. This is fairly large, and contains a lot of components we're not actively using. Given this, it's more efficient to implement only the individual JavaScript files—and their dependency utilities—that we need. This does introduce a small amoutn of overhead due to common functions potentially being repeated in each script, but that is trivial relative to the net savings.