Removed local copies of vendor stylesheets
We had three sets of vendor SASS files stored as part of the repository: Zurb Foundation, Sassy Lists, and Normalize. The latter two were only needed as dependencies of Zurb Foundation. But Zurb Foundation is already downloaded and maintained as an npm depndency—and, alongside the previous commit (131c400), most dependencies were already redirected to that copy. As a result, we can delete the 100+ files that were being unnecessarily duplicated as part of this repository.