Removed duplicate settings
The `_settings.scss` only needs to override variables that have changed; there is no need to duplicate existing settings. In fact, doing so actually makes the configuration harder to maintain since it's unclear if mismatches with the base `_settings.scss` are deliberate changes on our part, or updates on Zurb's part. Given this, I've carefully deduped the files, and removed any definitions that matched the base definition.

As part of this, I've updated the previous values for any where the values didn't match, and also flagged items that were either added since we last forked `_settings.scss` (in case we want to modify them) or which have been since removed (in case we need to update a different variable).