Removed newly introduced variables
There were a number of variables that were newly introduced in the source `_settings.scss`. These were likely variables that Foundation has added since we had originally forked the file. To be safe, I had flagged them until I had a chance to look into each.

Since none of these are used in our own `*.scss` files, I'm going to assume we're fine with the defaults. This commit, however, provides documentation of what was new in case we ever have a need to modify these values in the future.