Introduce support for handling Font Awesome via SASS
Currently, the Font Awesome icons are being handled via JavaScript, which replaces the associated classes with SVG icons dynamically. This is actually pretty fast, but the JavaScript for it is really large (over a megabyte for all fonts!). By using the inline version, we don't require the Font Awesome CSS nor the Font Awesome JavaScript—though we still need the fonts themselves, which are about 80k each. This commit establishes the framework for this, but doesn't replace any icons with SCSS implementations.