Swapped out `phone` and `email` icons for SCSS implementations
There were two instances of these—one in the contact pane of the homepage, and another in the footer—with slightly different treatment. As a result, this required some special handling. This isn't identical to the previous implementation, due to variations in the fonts, and how SVGs are handled, but it's really close, and should satisfy our requirements.