Removed unused library navigation elements
These were left over from an original concept for the library page which had iconography—mapping to the service sections—which could be used to filter the content, as well as second-tier navigation for further filtering by tags. We may reevaluate this in the future should make active use of the library. For now, however, there isn't enough content to justify this, and we'd need to work out the design regardless.