Removed Font Awesome JavaScript
The Font Awesome JavaScript is an alternative to using the Font Awesome fonts. It looks for the same classes associated with the Font Awesome CSS, and swaps them out dynamically with SVG files. This is actually surprisingly fast. But the JavaScript is really large—in fact, at over a megabyte, it's even larger than using all three fonts.

Since we've replaced all Font Awesome icons with SCSS mixins, we don't need either the JavaScript or the standard CSS.

In the future, we may want to take this a step further and replace Font Awesome entirely with inline SVG calls, since we really only need a handful of icons.