Refactored background images as tiles
The backgrounds may look like solid colors on initial glance, but they're actually very subtle textures. Those background images were fairly large—with the `Introduction.Tile.png` being over 80k! But these are actually just repeating tiles, which can be reduced down to that basic element. Doing so reduces these considerably. For instance, the new `Introduction.Background.Tile.png` is now only 6k.

As part of this, I established a naming convention of `{image}.Background.png` for the composite version—just in case we need to go back to it for some reason—and `{Image}.Background.Tile.png` for the tiling version.

This includes updating the CSS paths to ensure the new tiles are being correctly referenced.