Introduce lazy loading for off-screen images
This implementation relies on browser support, which is limited to Chrome and Edge right now. A more sophisticated approach will require the use of a polyfill to provide support for other (or older) browsers. That said, the built-in approach is a quick performance enhancement for a large share of our audience.