Preload local fonts and images
Without `preload`, the browser needs to first load the CSS file to discover the dependencies on the `Hero.Large.png` (which shows up above the fold) and the Font Awesome fonts. By adding the `<link rel="preload" />` directive, we tip the browser that these will need to be loaded, and give it the option of getting a headstart on them.