Improve download performance for Google Fonts
For now, I'm keeping the Google Fonts definition in the SCSS (d0cee65). In addition, however, I'm adding it as a direct stylesheet reference in the `_Layout.cshtml`. This allows the browser to immediately begin downloading these resources. This allows the browser to discover the fonts quicker, and start downloading them.

While I was at it, I also added `display=swap` so that the page can render while the fonts are still downloading, as well as a `dns-prefetch` directive for `fonts.gstatic.com` so that the browser can lookup the DNS information even before discovering the fonts, should it have the resources to do so.

(A `preconnect` would be even faster—but that would also be wastedful for anyone with those fonts already cached, and thus I'm using `dns-prefetch` instead.)