Introduced Google Analytics
In addition to the standard Google Analytics tag, this also includes a `dns-prefetch` for the script's `googletagmanager.com` hostname, as well as a `preconnect` for the actual `google-analytics.com` call. This should help ensure that the connection is ready as soon as the script is called, thus reducing the overhead for that connection. Though, it should be noted, as this isn't involved in rendering the web page, that's not as critical as other, similar optimizations.