Ensured that the `IgniaTopicViewModelLookupService` supports OnTopic Editor
In order to map view models associated with the OnTopic Editor, the `ITopicMappingService` must be given an `ITypeLookupService` which includes the OnTopic Editor's view models. This is handled by the `EditorViewModelLookupService`. The `IgniaTopicViewModelLookupService`, however, was deriving from teh `TopicViewModelLookupService`, which doesn't include the editor types. Easy fix.

(Note that the `EditorViewModelLookupService` itself derives from the `TopicViewModelLookupService`, so this retains the original registrations as well.)