Updated from GSAP 1.20.2 to 2.1.3
This is an easy, non-breaking change to minimize the variables when migrating to GSAP 3.0, as GSAP 1.20.5 and GSAP 2.0.0 are the same library. As such, this is actually a minor bump to the latest 1.x/2.x code, and not a major update.

(GSAC updated to a major version to avoid a breaking change with npm, due to how their module format changed, but the compiled/minified files do not have breaking changes.)