Updated from GSAP 2.1.3 to GSAP 3.5.1
Now that we're just doing a straight migration from GSAP 2.x to 3.x, we can follow the standard migration guide. This commit updates to the latest version, adjusts the build process to account for the new directory structure, and adjusts the file references to call `gsap.min.js` instead of `TweenMax.min.js`.

This doesn't work yet, likely due to missing plugins. I'll be working on those in the next couple of updates.