Added the `CSSRulePlugin` to address an `autoCSS` warning
The legacy code used an `autoCSS` property on calls to e.g. `staggerFromTo()` and `TweenMax.to()`. It's unclear if this is needed. According to the v2 documentation, this "allows CSSPlugin to automatically parse CSS properties from the vars :{} object when the target of a tween is a DOM element." To renable that support, I've explicitly introduced the `CSSRulePlugin` script, which silences the warning.

That said, we should evaluate if we're actually using any of these CSS plugins, as I believe they're less performant, and carry along this extra dependency.