Migrated to new easing syntax
I actually prefer the legacy syntax since it doesn't rely on string parsing. That said, I'm migrating to the new syntax to ensure our code is readable to people coming in at GSAP 3.x+ and to make it easier to evaluate against current samples (which mostly use the new syntax).