Migrated to new stagger syntax
Migrated from the legacy `TweenMax.staggerFromTo()` syntax to the new `gsap.fromTo()` method. The special attributes—such as `duration` and `stagger`—are moved to the `toVars` object only; they aren't needed on the `fromVars` object.

It's worth noting that the animation and staggering is slightly different here than it was in the origina 1.20.2 version of the code. That appears to be due to subtle differences in the underlying libraries, as this was true even before changing the configuration. The changes aren't significant, and only show up within the clients section, which displays the content a bit faster. This should be acceptable for our needs, but worth being aware of, in case we want to tweek the configuration to account for this.